see notes
					 /\
added in ws node server is runnign as so /\

can now iterate state is passed correctly between python and node.
we get about 300 iterations ps when just cycling through and randomly choosing an action

have also allowed actions to be passed across and implemented a new model on the right___________>>>>>>>>>>>>

Just with random choices, it got like 6-8 on average, however it now prints its average.

please dont close it!! :)

The link to the place I stole it is here  \/
https://towardsdatascience.com/reinforcement-learning-w-keras-openai-dqns-1eed3a5338c

in the top right we have a deep q agent training. it takes about 100 iterations per second,
but only 2 of that is gathering data, the rest is training. (it takes 2 sec to play a game ?!?!?!!?!?!)

1. We can still adjust the training to make it only traing on the memories where it does better than the average...

2. I assume we can improve the efficiency of the training. I think i can batch up the experiences and then train on on one huge thing

3. we need to be able to train models better :((